changed the features as required
