+++ package.json
@@ -2,6 +2,7 @@
   "name": "mentors-frontend",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://192.168.1.125:3000",
   "dependencies": {
     "axios": "^0.18.0",
     "history": "^4.9.0",
