+++ src/App.js
@@ -1,17 +1,12 @@
 import React from 'react';
-import { Switch, Route } from "react-router-dom";
 import './App.css';
-import SignUp from './components/SignUp';
-import Landing from './components/Landing';
+import Authentication from './components/Authentications/Authentication';
 
 function App() {
     return (
-        <React.Fragment>
-            <Switch>
-                <Route exact path='/' component={Landing} />
-                <Route path='/signin' component={SignUp} />
-            </Switch>
-        </React.Fragment>
+        <>
+            <Authentication />
+        </>
     );
 }
