+++ src/components/Authentications/Authform.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import Login from './Login'
+import SignUp from './SignUp'
+
+const AuthForm = () => {
+    return (
+        <>
+            <Login />
+            <SignUp />
+        </>
+    )
+}
+
+export default AuthForm
