--- src/components/Landing.jsx
@@ -1,7 +1,0 @@
-import React, { Component } from "react";
-
-export default class Landing extends Component {
-  render() {
-    return <div>Landing</div>;
-  }
-}
